+++ src/render/painter.ts
@@ -79,7 +79,7 @@ type PainterOptions = {
 /**
  * Initialize a new painter object.
  *
- * @param {Canvas} gl an experimental-webgl drawing context
+ * @param {Canvas} gl a webgl drawing context
  * @private
  */
 class Painter {
@@ -511,7 +511,10 @@ class Painter {
 
     gpuTimingStart(layer: StyleLayer) {
         if (!this.options.gpuTiming) return;
-        const ext = this.context.extTimerQuery;
+        const timing = this.context.timing;
+        if (!timing)
+            return;
+
         // This tries to time the draw call itself, but note that the cost for drawing a layer
         // may be dominated by the cost of uploading vertices to the GPU.
         // To instrument that, we'd need to pass the layerTimers object down into the bucket
@@ -521,17 +524,17 @@ class Painter {
             layerTimer = this.gpuTimers[layer.id] = {
                 calls: 0,
                 cpuTime: 0,
-                query: ext.createQueryEXT()
+                query: timing.createQuery()
             };
         }
         layerTimer.calls++;
-        ext.beginQueryEXT(ext.TIME_ELAPSED_EXT, layerTimer.query);
+        timing.beginQuery(timing.TIME_ELAPSED, layerTimer.query);
     }
 
     gpuTimingEnd() {
         if (!this.options.gpuTiming) return;
-        const ext = this.context.extTimerQuery;
-        ext.endQueryEXT(ext.TIME_ELAPSED_EXT);
+        const timing = this.context.timing;
+        timing.endQuery(timing.TIME_ELAPSED);
     }
 
     collectGpuTimers() {
@@ -542,11 +545,12 @@ class Painter {
 
     queryGpuTimers(gpuTimers: {[_: string]: any}) {
         const layers = {};
+        const timing = this.context.timing;
         for (const layerId in gpuTimers) {
             const gpuTimer = gpuTimers[layerId];
-            const ext = this.context.extTimerQuery;
-            const gpuTime = ext.getQueryObjectEXT(gpuTimer.query, ext.QUERY_RESULT_EXT) / (1000 * 1000);
-            ext.deleteQueryEXT(gpuTimer.query);
+
+            const gpuTime = timing.getQueryParameter(gpuTimer.query, timing.QUERY_RESULT) / (1000 * 1000);
+            timing.deleteQuery(gpuTimer.query);
             layers[layerId] = gpuTime;
         }
         return layers;